Implemented possibility to give geometry in bohr
Modified get_geometry_input_file to check if units keyword is in the
geometry block and return a logical on whether or not it's Ångström.

Modified read_geometry_molecular_system to multiply each coordinate by a
conversion factor; if Bohr was specified, the factor is
bohr_to_angstrom. If Ångström was specified, or no units at all were
specified, the factor is 1.

Modified move_to_section_input_file to check if the requested section is
geometry, and if so, to not count the units-line as a record if it is
present.

Modified print_system_molecular_system to print the units specified.
Default is Ångström.